fix: Move api from server to the individual modules
Packages can be self sufficient if we move basic types inside them.
This is going to give us couple major advantages:

- Packages are decoupled from the server and can be used with plain Apollo if needed
- Each package contains default implementation that we support
- Single responsibility in one place
- Better aproach for documenting
- Simple dependency model (server requires individual packages)
- Makes server only and integration/getting started experience package
- It makes things simply to understand and contribute
