+++ packages/voyager-audit/src/api/AuditLogger.ts
@@ -1,0 +1,6 @@
+import { GraphQLResolveInfo } from 'graphql'
+
+export interface AuditLogger {
+  logResolverCompletion (msg: string, success: boolean, obj: any, args: any, context: any, info: GraphQLResolveInfo): void
+  auditLog (msg: string, obj: any, args: any, context: any, info: GraphQLResolveInfo): void
+}
