+++ packages/voyager-audit/src/api/index.ts
@@ -1,0 +1,2 @@
+export * from './AuditLogger'
+export * from './DefaultAuditLogger'
