+++ packages/voyager-audit/src/index.ts
@@ -1,1 +1,2 @@
 export * from './auditLog'
+export * from './api'
