+++ packages/voyager-keycloak/src/AuthContextProvider.ts
@@ -1,6 +1,8 @@
 import { AuthContextProvider } from '@aerogear/voyager-server'
 import { Request } from 'express'
 import { IncomingMessage } from 'http'
+import { AuthContextProvider } from './api'
+
 export class KeycloakAuthContextProvider implements AuthContextProvider {
   public readonly request: any
   public readonly accessToken: any
