+++ packages/voyager-keycloak/src/KeycloakSecurityService.ts
@@ -2,6 +2,7 @@ import { SecurityService } from '@aerogear/voyager-server'
 import { Router } from 'express'
 import session from 'express-session'
 import Keycloak from 'keycloak-connect'
+import { SecurityService } from './api'
 import { KeycloakAuthContextProvider } from './AuthContextProvider'
 import { schemaDirectives } from './schemaDirectives'
