+++ packages/voyager-keycloak/src/api/AuthContextProvider.ts
@@ -1,0 +1,5 @@
+
+export interface AuthContextProvider {
+  isAuthenticated (): boolean
+  hasRole (role: string): boolean
+}
