+++ packages/voyager-keycloak/src/index.ts
@@ -1,1 +1,2 @@
 export * from './KeycloakSecurityService'
+export * from './api'
