+++ packages/voyager-metrics/src/api/Metrics.ts
@@ -1,0 +1,3 @@
+export interface Metrics {
+  updateResolverMetrics (resolverInfo: any, responseTime: number): void
+}
